Add golden tests for milestone examples
