+++ test/BackendC.hs
@@ -1,7 +1,8 @@
 module BackendC where
 
 import BackendC.Positive qualified as P
+import BackendC.Examples qualified as E
 import Base
 
 allTests :: TestTree
-allTests = testGroup "Backend C tests" [P.allTests]
+allTests = testGroup "Backend C tests" [P.allTests, E.allTests]
