+++ test/BackendC/Positive.hs
@@ -1,12 +1,7 @@
 module BackendC.Positive where
 
+import BackendC.Base
 import Base
-import Data.FileEmbed
-import Data.Text.IO qualified as TIO
-import MiniJuvix.Pipeline
-import MiniJuvix.Translation.MonoJuvixToMiniC as MiniC
-import System.IO.Extra (withTempDir)
-import System.Process qualified as P
 
 data PosTest = PosTest
   { _name :: String,
@@ -30,88 +25,8 @@ testDescr PosTest {..} =
    in TestDescr
         { _testName = _name,
           _testRoot = tRoot,
-          _testAssertion = Steps $ \step -> do
-            step "Check clang and wasmer are on path"
-            assertCmdExists "clang"
-            assertCmdExists "wasmer"
-
-            step "Lookup WASI_SYSROOT_PATH"
-            sysrootPath <-
-              assertEnvVar
-                "Env var WASI_SYSROOT_PATH missing. Set to the location of the wasi-clib sysroot"
-                "WASI_SYSROOT_PATH"
-
-            step "C Generation"
-            let entryPoint = defaultEntryPoint mainFile
-            p :: MiniC.MiniCResult <- runIO (upToMiniC entryPoint)
-
-            expected <- TIO.readFile expectedFile
-
-            step "Compile C with standalone runtime"
-            actualStandalone <- clangCompile (standaloneArgs sysrootPath) p step
-            step "Compare expected and actual program output"
-            assertEqDiff ("check: WASM output = " <> expectedFile) actualStandalone expected
-
-            step "Compile C with libc runtime"
-            actualLibc <- clangCompile (libcArgs sysrootPath) p step
-            step "Compare expected and actual program output"
-            assertEqDiff ("check: WASM output = " <> expectedFile) actualLibc expected
+          _testAssertion = Steps $ clangAssertion mainFile expectedFile ""
         }
-
-clangCompile ::
-  (FilePath -> FilePath -> [String]) ->
-  MiniC.MiniCResult ->
-  (String -> IO ()) ->
-  IO Text
-clangCompile mkClangArgs cResult step =
-  withTempDir
-    ( \dirPath -> do
-        let cOutputFile = dirPath </> "out.c"
-            wasmOutputFile = dirPath </> "out.wasm"
-        TIO.writeFile cOutputFile (cResult ^. MiniC.resultCCode)
-        step "WASM generation"
-        P.callProcess
-          "clang"
-          (mkClangArgs wasmOutputFile cOutputFile)
-        step "WASM execution"
-        pack <$> P.readProcess "wasmer" [wasmOutputFile] ""
-    )
-
-standaloneArgs :: FilePath -> FilePath -> FilePath -> [String]
-standaloneArgs sysrootPath wasmOutputFile cOutputFile =
-  [ "-nodefaultlibs",
-    "-I",
-    takeDirectory minicRuntime,
-    "--target=wasm32-wasi",
-    "--sysroot",
-    sysrootPath,
-    "-o",
-    wasmOutputFile,
-    wallocPath,
-    cOutputFile
-  ]
-  where
-    minicRuntime :: FilePath
-    minicRuntime = $(makeRelativeToProject "minic-runtime/standalone/minic-runtime.h" >>= strToExp)
-    wallocPath :: FilePath
-    wallocPath = $(makeRelativeToProject "minic-runtime/standalone/walloc.c" >>= strToExp)
-
-libcArgs :: FilePath -> FilePath -> FilePath -> [String]
-libcArgs sysrootPath wasmOutputFile cOutputFile =
-  [ "-nodefaultlibs",
-    "-I",
-    takeDirectory minicRuntime,
-    "-lc",
-    "--target=wasm32-wasi",
-    "--sysroot",
-    sysrootPath,
-    "-o",
-    wasmOutputFile,
-    cOutputFile
-  ]
-  where
-    minicRuntime :: FilePath
-    minicRuntime = $(makeRelativeToProject "minic-runtime/libc/minic-runtime.h" >>= strToExp)
 
 allTests :: TestTree
 allTests =
