+++ tests/examplesExpected/MiniTicTacToe/expected.golden
@@ -1,0 +1,69 @@
+MiniTicTacToe
+-------------
+
+Type a number then ENTER to make a move
+
++---+---+---+
+| 1 | 2 | 3 |
++---+---+---+
+| 4 | 5 | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player X: 
++---+---+---+
+| X | 2 | 3 |
++---+---+---+
+| 4 | 5 | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player O: 
++---+---+---+
+| X | O | 3 |
++---+---+---+
+| 4 | 5 | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player X: 
++---+---+---+
+| X | O | X |
++---+---+---+
+| 4 | 5 | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player O: 
++---+---+---+
+| X | O | X |
++---+---+---+
+| O | 5 | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player X: 
++---+---+---+
+| X | O | X |
++---+---+---+
+| O | X | 6 |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player O: 
++---+---+---+
+| X | O | X |
++---+---+---+
+| O | X | O |
++---+---+---+
+| 7 | 8 | 9 |
++---+---+---+
+Player X: 
++---+---+---+
+| X | O | X |
++---+---+---+
+| O | X | O |
++---+---+---+
+| X | 8 | 9 |
++---+---+---+
+Player X wins!
