+++ tests/examplesExpected/ValidityPredicates/expected.golden
@@ -1,0 +1,1 @@
+VP Status: OK
